+++ src/App.test.js
@@ -3,6 +3,6 @@ import App from './App';
 
 test('renders learn react link', () => {
   render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
+  const linkElement = screen.getByText(/0/i);
   expect(linkElement).toBeInTheDocument();
 });
