+++ src/components/Calculator.js
@@ -1,19 +1,11 @@
-import PropTypes from 'prop-types';
 import ResultScreen from './ResultScreen';
 import Keypad from './Keypad';
 
-const Calculator = (props) => {
-  const { keypad } = props;
-  return (
-    <div className="calculator-wrapper">
-      <ResultScreen />
-      <Keypad keypad={keypad} />
-    </div>
-  );
-};
-
-Calculator.propTypes = {
-  keypad: PropTypes.arrayOf.isRequired,
-};
+const Calculator = () => (
+  <div className="calculator-wrapper">
+    <ResultScreen />
+    <Keypad />
+  </div>
+);
 
 export default Calculator;
