+++ src/components/KeyBtn.js
@@ -2,9 +2,11 @@ import PropTypes from 'prop-types';
 
 const KeyBtn = (props) => {
   const { button } = props;
-  const classes = `${button.category} ${button.value === '0' && 'double'}`;
+  const classes = button.value === '0' ? `${button.category} keybtn double` : `${button.category} keybtn`;
   return (
-    <div className={classes}>{button.value}</div>
+    <div className={classes}>
+      <span>{button.value}</span>
+    </div>
   );
 };
